disable button on call

+++ src/Storefront/Controller/IdealFastCheckoutController.php
@@ -77,7 +77,16 @@ class IdealFastCheckoutController extends AbstractPaymentController
                 ])
             );
 
+            if ($redirectPath == null) {
+                return $this->response([
+                    "status" => "FAILED",
+                    "message" => "Something went wrong",
+                    "reload" => true
+                ], true);
+            }
+
             $this->cartService->deleteFromCart($salesChannelContext);
+
             return $this->response([
                 "redirect" => $redirectPath
             ]);
